+++ packages/eyes-api/CHANGELOG.md
@@ -5,6 +5,9 @@
 - add `iPad_Air_4` item with value `'iPad Air (4th generation)'` to `IosDeviceName` enum
 - add `Galaxy_Tab_S7` item with value `'Galaxy Tab S7'` to `DeviceName` enum
 - add patch for new default default behavior of the core
+- updated to @applitools/logger@1.0.9 (from 1.0.5)
+- updated to @applitools/types@1.0.24 (from 1.0.19)
+- updated to @applitools/utils@1.2.11 (from 1.2.4)
 
 ## 1.1.6 - 2021/11/10